refactor(text-processing): merge sliding window chunker into main chunkers module
- Move SlidingWindowSemanticSplitter class from separate file into chunkers.py
- Update imports in test files to reflect new class location
- Update docstrings to document the integration
- Fix test_sliding_window_chunker.py to import from chunkers instead of the now-deleted module
- Verify functionality with tests and CLI commands

This change simplifies the codebase by consolidating related chunking functionality
in a single module, making maintenance easier while preserving all existing
functionality.
