+++ tests/test_sliding_window_chunker.py
@@ -12,9 +12,7 @@ import string
 from sentence_transformers import SentenceTransformer
 from phentrieve.config import get_sliding_window_config_with_params
 from phentrieve.text_processing.pipeline import TextProcessingPipeline
-from phentrieve.text_processing.sliding_window_chunker import (
-    SlidingWindowSemanticSplitter,
-)
+from phentrieve.text_processing.chunkers import SlidingWindowSemanticSplitter
 
 
 class TestSlidingWindowSplitter(unittest.TestCase):
